+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Moonrise compatibility
+
 ### Fixed
 - Containers remaining open by vanished players in certain conditions
 - Pressure plates remaining pressed by vanished players in certain conditions